+++ src/main/java/com/yh/web/dto/board/Board.java
@@ -6,15 +6,15 @@ import java.time.LocalDateTime;
 
 @Data
 public class Board {
-    private int articleNo;
+    private long articleNo;
     private String title;
     private String content;
     private String writer;
     private LocalDateTime regDate;
-    private int recommend;
-    private int hit;
-    private int parent;
-    private int grpNo;
+    private long recommend;
+    private long hit;
+    private long parent;
+    private long grpNo;
     private boolean pub;
     private String ip;
 }
